backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
psycopg2-binary==2.9.9
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1
pgvector==0.2.4
django-celery-results==2.5.1
django-environ==0.11.2
pytest==8.1.1
pytest-django==4.8.0
gunicorn==21.2.0